Finish refactoring book cover methods, and move tests
